+++ KAkica.Communication/Request/OwnerRequest.cs
@@ -5,14 +5,12 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace KAkica.API.Request
+namespace KAkica.Communication.Request
 {
-    public class PooperRequest : JokJaBreRequest<Pooper>
+    public class OwnerRequest : JokJaBreRequest<Owner>
     {
-        public string Name { get; set; }
-        public override void SetTo(Pooper obj)
+        public override void SetTo(Owner obj)
         {
-
         }
     }
 }
