+++ KAkica.Communication/Request/PooperRequest.cs
@@ -5,12 +5,14 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace KAkica.API.Request
+namespace KAkica.Communication.Request
 {
-    public class OwnerRequest : JokJaBreRequest<Owner>
+    public class PooperRequest : JokJaBreRequest<Pooper>
     {
-        public override void SetTo(Owner obj)
+        public string Name { get; set; }
+        public override void SetTo(Pooper obj)
         {
+
         }
     }
 }
