+++ KAkica_API.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KAkica.API", "KAkica_API\KA
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KAkica.Domain", "KAkica.Domain\KAkica.Domain.csproj", "{FFEF7DA2-35AB-4732-8388-42461B4797CA}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KAkica.Communication", "KAkica.Communication\KAkica.Communication.csproj", "{A8B1CA2A-65B4-4243-AF5B-5A5B41444ADD}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Frontend", "Frontend", "{D83F62B1-97C6-4A84-9894-AA3A45958920}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,10 @@ Global
 		{FFEF7DA2-35AB-4732-8388-42461B4797CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FFEF7DA2-35AB-4732-8388-42461B4797CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FFEF7DA2-35AB-4732-8388-42461B4797CA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8B1CA2A-65B4-4243-AF5B-5A5B41444ADD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8B1CA2A-65B4-4243-AF5B-5A5B41444ADD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8B1CA2A-65B4-4243-AF5B-5A5B41444ADD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8B1CA2A-65B4-4243-AF5B-5A5B41444ADD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
